requirements.txt
gremlinpython==3.7.1
boto3==1.34.69
python-dotenv==1.0.1
aiohttp==3.9.1
async_timeout==5.0.1